Fix: export addMartechStack to geo-script
